On branch gh-299-database-password-validation-relaxation
https://github.com/wls-eng/arm-oraclelinux-wls/issues/299

modified:   arm-oraclelinux-wls-cluster/pom.xml

- Reduce `PasswordMinLength` to 5.

modified:   arm-oraclelinux-wls-cluster/src/main/arm/createUiDefinition.json

The validation requirements for the database password field must only be as strict as the least strict of the three supported databases. According to Oracle, that means the validation requirements should be 5 chars min, at least one number.

This issue asks the assignee to update the createUiDefinition.json so that

- The validation regEx represents that requirement.
- The validation message states that requirement.

Signed-off-by: Ed Burns <edburns@microsoft.com>